📜 Scribe: Documentation update for Normalized Assignment Model
- Relocated `Quiz` and `Homework` data classes to dedicated files for better visibility.
- Added comprehensive KDoc to normalized assessment entities and their templates.
- Documented the architectural distinction between `QuizDao`/`HomeworkDao` and their legacy counterparts.
- Expanded `data/README.md` with a "Comparison: Log Model vs. Normalized Model" guide.

This update clarifies the pedagogical and technical intent behind the application's evolving assessment data strategy without altering executable logic.
